requirements.txt
openai
googletrans==4.0.0-rc1
httpx==0.13.3
httpcore==0.9.1
deepl
requests
python-dotenv
deep-translator
cohere
elevenlabs
simpleaudio
pydub
ffmpeg